logging in or signing up directs a user to their own list. Showing all recipes in a users list, and showing the ingredients for all the recipes. The ingredients are together if more than one recipe has the same ingredient so there are no duplicates

+++ shopping_helper/db/migrate/20160703235238_create_ingredient_recipes.rb
@@ -1,7 +1,7 @@
 class CreateIngredientRecipes < ActiveRecord::Migration
   def change
     create_table :ingredient_recipes do |t|
-      t.float :quantity
+      t.decimal :quantity
       t.string :quantity_type
       t.references :recipe, index: true, foreign_key: true
       t.references :ingredient, index: true, foreign_key: true
